+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # [Unreleased]
 
 - Add Enum::instanceFromKey($key)
+- **Breaking** Change `$instance->identifier` to `$instance->name()`
+- **Breaking** Change `Enum::identifiers()` to `Enum::names()`
+- **Breaking** Change `Enum::getKeyForIdentfier()` to `Enum::keyForName()`
 
 # 1.1.0